Improve client API performance
Things are really slow right now because we have a separate Cloud Function for each endpoint. At scale, that's fine in theory because those functions will always be running. However, since we don't have a lot of usage currently, each function has to boot up and it takes a few seconds. This CL merges all the functions into one so that function instances can be reused.

Signed-off-by: Alex Saveau <alexsaveau@bymason.com>